+++ internal/pkg/server/server.go
@@ -66,6 +66,7 @@ type Server struct {
 	// within that Git folder.
 	analyzedFiles map[string]map[string]bool
 
+	composeSupport    bool
 	composeCompletion bool
 
 	mutex sync.RWMutex
@@ -87,6 +88,7 @@ func NewServer(docManager *document.Manager) *Server {
 		telemetry:                  telemetry.NewClient(),
 		scoutService:               scoutService,
 		sessionTelemetryProperties: sessionTelemetryProperties,
+		composeSupport:             true,
 		composeCompletion:          true,
 		diagnosticsCollectors: []textdocument.DiagnosticsCollector{
 			buildkit.NewBuildKitDiagnosticsCollector(),